Created addToCart function and called it in both POST and PUT payment routes due to similar code in both
