+++ src/app/api/payment/route.js
@@ -1,60 +1,16 @@
 import Stripe from 'stripe'
+import addToCart from '../../../../lib/addToCart'
 
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY)
-const orderStore = {}
 
 export async function POST(req) {
-    try {
-        const { amount, items } = await req.json()
-
-        const paymentIntent = await stripe.paymentIntents.create({
-            amount,
-            currency: 'usd',
-            automatic_payment_methods: { enabled: true },
-            metadata: { items: JSON.stringify(items) }
-        })
-
-        orderStore[paymentIntent.id] = { items, amount }
-
-        //Code to empty the cart
-
-        return new Response(JSON.stringify({
-            clientSecret: paymentIntent.client_secret,
-            paymentIntentId: paymentIntent.id
-        }), {
-            status: 200,
-            headers: { 'Content-Type': 'application/json' }
-        })
-    } catch (error) {
-        return new Response(JSON.stringify({ error: error.message }), {
-            status: 500,
-            headers: { 'Content-Type': 'application/json' }
-        })
-    }
+    const { amount, items } = await req.json()
+    return addToCart(amount, items)
 }
 
 export async function PUT(req) {
-    try {
-        const { amount, items, payment_intent_id } = await req.json()
-
-        const paymentIntent = await stripe.paymentIntents.update(payment_intent_id, {
-            amount,
-            metadata: { items: JSON.stringify(items) }
-        })
-
-        return new Response(JSON.stringify({
-            clientSecret: paymentIntent.client_secret,
-            paymentIntentId: paymentIntent.id
-        }), {
-            status: 200,
-            headers: { 'Content-Type': 'application/json' }
-        })
-    } catch (err) {
-        return new Response(JSON.stringify({ error: err.message }), {
-            status: 500,
-            headers: { 'Content-Type': 'application/json' }
-        })
-    }
+    const { amount, items, payment_intent_id } = await req.json()
+    return addToCart(amount, items, payment_intent_id)
 }
 
 export async function GET(req) {
@@ -64,13 +20,12 @@ export async function GET(req) {
     try {
         const paymentIntent = await stripe.paymentIntents.retrieve(payment_intent)
 
-        let items = []
-        items = JSON.parse(paymentIntent.metadata?.items || '[]')
+        //Code to empty the cart
 
         return new Response(JSON.stringify({
             amount: paymentIntent.amount,
             status: paymentIntent.status,
-            items
+            items: paymentIntent.metadata?.items ? JSON.parse(paymentIntent.metadata.items) : []
         }), {
             status: 200,
             headers: { 'Content-Type': 'application/json' }
